Restores on-join announcements for all jobs, with a twist - Majority of jobs are now announced on respective departmental channel, instead of public channel. - During red alert, all jobs are announced on main channel regardless of the above. This lets people know when possibly required emergency personnel arrives when situation is seriously screwed up already. - Vote: https://baystation12.net/forums/threads/vote-restore-on-join-announcements-for-non-command-crew.5091/ - Vote results: 16% for full revert (all jobs announced on main frequency), 60% this (departmental frequencies for most jobs), 24% no change(only few select jobs announced) - Bonus: Partially/Mostly fixes #18094 - resolves qdel loops caused by overmap edges.

+++ code/modules/overmap/spacetravel.dm
@@ -56,7 +56,8 @@ proc/overmap_spacetravel(var/turf/space/T, var/atom/movable/A)
 		return
 
 	if(A.lost_in_space())
-		qdel(A)
+		if(!QDELETED(A))
+			qdel(A)
 		return
 
 	var/nx = 1
